+++ R/dataframe__frame.R
@@ -167,7 +167,7 @@ pl$DataFrame = function(..., make_names_unique = TRUE, schema = NULL) {
       if (make_names_unique) {
         keys = make.unique(keys, sep = "_")
       } else {
-        stopf(
+        stop(
           paste(
             "conflicting column names not allowed:",
             paste(unique(keys[duplicated(keys)]), collapse = ", ")
@@ -487,10 +487,10 @@ DataFrame_schema = method_as_property(function() {
 
 #
 DataFrameCompareToOtherDF = function(self, other, op) {
-  stopf("not done yet")
+  stop("not done yet")
   #    """Compare a DataFrame with another DataFrame."""
-  if (!identical(self$columns, other$columns)) stopf("DataFrame columns do not match")
-  if (!identical(self$shape, other$shape)) stopf("DataFrame dimensions do not match")
+  if (!identical(self$columns, other$columns)) stop("DataFrame columns do not match")
+  if (!identical(self$shape, other$shape)) stop("DataFrame dimensions do not match")
 
   suffix = "__POLARS_CMP_OTHER"
   other_renamed = other$select(pl$all()$suffix(suffix))
@@ -1640,12 +1640,12 @@ DataFrame_glimpse = function(..., return_as_string = FALSE) {
 #' # explode a single column, append others
 #' df$explode("numbers")
 #'
-#' # it doesn't change anything if the input is not a list-column
+#' # it is also possible to explode a character column to have one letter per row
 #' df$explode("letters")
 #'
 #' # explode two columns of same nesting structure, by names or the common dtype
 #' # "List(Float64)"
-#' df$explode(c("numbers", "numbers_2"))
+#' df$explode("numbers", "numbers_2")
 #' df$explode(pl$col(pl$List(pl$Float64)))
 DataFrame_explode = function(...) {
   self$lazy()$explode(...)$collect()