+++ src/rust/src/rdataframe/mod.rs
@@ -3,6 +3,7 @@ use polars::prelude::{self as pl, IntoLazy, SerWriter};
 use std::result::Result;
 pub mod read_csv;
 pub mod read_ipc;
+pub mod read_ndjson;
 pub mod read_parquet;
 use crate::conversion_r_to_s::robjname2series;
 use crate::lazy;
@@ -528,6 +529,7 @@ extendr_module! {
     mod rdataframe;
     use read_csv;
     use read_ipc;
+    use read_ndjson;
     use read_parquet;
     use rdatatype;