+++ NEWS.md
@@ -26,8 +26,7 @@
 - `$scan_csv()` and `$read_csv()` now match more closely the Python-Polars API (#455):
   - `sep` is renamed `separator`, `overwrite_dtypes` is renamed `dtypes`, 
     `parse_dates` is renamed `try_parse_dates`.
-  - new arguments `missing_utf8_is_empty_string`, `rechunk`, `eol_char`, 
-    `raise_if_empty`, `truncate_ragged_lines`
+  - new arguments `rechunk`, `eol_char`, `raise_if_empty`, `truncate_ragged_lines`
   - `path` can now be a vector of characters indicating several paths to CSV files. 
     This only works if all CSV files have the same schema.